habitaclia_project/requirements.txt
# Core scraping dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
lxml>=4.9.3
numpy>=1.24.0

# AI/ML Stack para embeddings
openai>=1.0.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
python-dotenv>=1.0.0

# Optional analytics
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Development tools
jupyter>=1.0.0
ipykernel>=6.25.0